Remove setTimeout, changed anchor tag to Link componend and renamed files in titlecase
